Added skeleton for Signup Page
