+++ src/App.tsx
@@ -2,6 +2,7 @@ import { Routes, Route, Navigate } from 'react-router-dom';
 
 import LoginPage from './pages/auth/LoginPage';
 import HomePage from './pages/home/HomePage';
+import SignUpPage from './pages/auth/SignupPage';
 
 import { Toaster } from 'react-hot-toast';
 
@@ -11,6 +12,7 @@ const App = () => {
       <Routes>
         <Route path="/" element={<Navigate to="/login" />} />
 				<Route path='/login' element={<LoginPage />} />
+				<Route path='/signup' element={<SignUpPage/>} />
 			</Routes>
       <Toaster/>
     </div>
